backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
phonenumbers==9.0.13
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
